io.metadata: ResultSetHdfLoader and write_dataset
PerDetectorHdf5 is now deprecated, but left in place as a shell
(probably not for long).  The ResultSetHdfLoader is tuned for the
context/metadata system, with a single API (the one hacked in through
PerDetectorHdf5.loader_class).  For writing datasets, a function makes
more sense than having a write_dataset method, assuming we don't want
to put HDF5 i/o right in the ResultSet class.

Also added a test for the array string conversion function.