Support cuts in coords/pmat.py. Small rewrite in get_footprint for clarity.
Handling cuts in the pointing matrix this way isn't enough to suppor
them in the maximum-likelihood framework, but it makes the rest of the
implementation a bit cleaner. However, right now enabling pmat cuts
leads to some unexpected small changes in the result, which should be
investigated before they are used properly.\n\nThe get_footprint change
was made while I was debugging something else. What was there before was
fine too, but I left these changes in because I think it makes it a bit
easier to see the logic. While list comprehensions are very explicit,
their verbosity can make it hard to see the forest for all the trees.