Add co-sampled encoder data to timestream
To enable bundling with the SMuRF data, the HK encoder data must be
co-sampled (interpolated) to match time stamps.

In this simple implementation, we use a linear interpolation
provided by np.interp(). Samples beyond the time limits of the
frame (for the first and last frames) are filled with NaNs rather
than extrapolated.