Fix bug that returns incomplete Books
The nature of the bug is that it would handle one azimuth turn-
around per HK frame added, i.e., if there are 10 HK frames in
total, only the first 10 turnarounds get processed and written to
the output Book. This is due to erroneously placed return
statements.

Also added a new `flush` method to reduce repeated code.