[Major] Restructure Bookbinder object and main loop
Based on actual functionality, the main loop is now the Bookbinder
object; the previous Bookbinder object is renamed the
FrameProcessor. Previous global variables are also moved into the
(new) Bookbinder object, which allows them to be updated as needed.

This update also eliminates the use of the G3Pipeline, which has
proven very limiting with its rigid serial paradigm.