Remove repeated turnaround check
Previously, the check for azimuth turnarounds was happening twice
for each frame added: first time when checking FrameProcessor
.ready(), second time when computing turnaround times. As the
turnaround finder is expected to grow more complex, this is an
unnecessary increase in computational time. This commit removes the
repetition: the turnaround times are computed once and stored as an
attribute of _HKBundle that can be retrieved on demand.