Change how long frames are handled
The solution to excessively long frames in #2 was to check the
length of the frame when rebundling, and split the output frame if
it would exceed the maximum specified length. However, there is a
possibility that the buffered frame would take up too much memory
if too many SMuRF frames are added before rebundling is triggered.
This commit adds a check for buffer length every time a new SMuRF
frame is added, and outputs a frame when the maximum specified
length is exceeded.

This bypasses (and replaces) the original solution of #2.