metadata: change implementation of SuperLoader
There's a lot of moving stuff around here but the main point of it is
to pass around the "det_info" ResultSet for metadata processing of
dets:* fields instead of the DetDb.  Subsequent to this we will be
able to have metadata lines modify the det_info, and also have the
det_info returned in the AxisManager on get_obs or get_meta.

References to ResultSet have been removed from AxisManager, as has the
restrict_dets method.