Refactor default mode
A lot of the original 'default mode' block is repeated code, which
increases the work to maintain the code for no reason. The solution
is to set a frame split 'event' in the distant future, and let the
FrameProcessor return max-length frames (which was also the length
used by the now-removed 'default mode').