New feature: Divide files based on total samples
A new output SMuRF file should be created once the current output
file has exceeded a certain size. Here, we take total number of
samples (including all channels) as a proxy for the file size; for
now, we use 1e9 samples as the criterion.

To do: ensure that all output files for each wafer are divided in
the same place.