Account for different number of channels
Because each wafer will have a different number of channels, we
need to ensure that the files with the largest number of channels
do not exceed the maximum size. In this commit, we add a keyword
argument to the Bookbinder denoting the largest number of channels
in the Book, which will be used to determine where to split the
output files.

The largest number of channels in the Book is determined by the
Bookbind routine in the SO-Workflow repo, just before calling
Bookbinder.