Address flush_time issue with zero samples
If there are no samples in the SmurfBundle (for whatever reason),
there is no 'last sample' to access, which leads to an error. Since
the DEFAULT_TIME attribute has been added, it can be used instead
to achieve the same goal of providing a time beyond the final
sample in the Book.