Fix bug: final flush time is past end of HK data
The velocity vector is 1 sample shorter than the position vector,
which causes problems with trying to rebundle the HKBundle at the
very end.

Also added the specific AttributeError exception since that should
only be caught in the case where there is no az/el data in the HK
file (HKBundle will be NoneType in that case).