Handle gaps in between SMuRF data frames
This scenario is more complicated the HK gaps since dropped frames
typically only occur on one wafer and not the others, yet the
Bookbinder cannot tell on its own whether the gap occurs in ALL
wafers or only a subset of them. This poses a problem since the HK
output (A-file) needs to be co-sampled against the SMuRF samples.

If there are gaps in the SMuRF data, they need to be flagged and
assigned an obviously wrong value (-1). The gap will be filled in
using either:
1. An estimation of the missing samples based on the (approximate)
   sampling interval in the latest frame, or
2. A user-supplied list of timestamps in the Book, perhaps obtained
   from the original database.
In Case 1, in the event that the gap only exists in the current
wafer, the fill-in samples will not line up with the correct
samples in the other wafers. It is therefore recommended that the
user compile a list of timestamps prior to running (the current
version of) the Bookbinder.

Also create a test for these two cases.