Bookbinder: Calculate true timestamps if available
Timestamps recorded in smurf_frame['data'].times are when the data
are first received by the smurf-streamer, and the true timestamps
must be reconstructed from the timing counters in the smurf_frame[
'primary'] timestream.

The logic for calculating the timestamps is copied from load_smurf
.py (Jan 23, 2023).

If the 'primary' timestream is unavailable, or the counters are all
zero, then fall back to the smurf_frame['data'].times timestamps.

Remove redundancies in existing functions, and create relevant unit
tests.