Fix 0-length buffer bug in Bookbinder
If the buffer is exactly maxlength, the Bookbinder clears the
buffer without advancing the frame split time and gets stuck. This
fix changes the >= maxlength to a strict >. The split time written
is then changed to be the first sample of the next frame,
consistent with the convention for HK-determined splits.