coords: make maps of demodQ/demodU properly
I believe the former code was only correct in the limit where the
parallactic angle on the sky was ~0; that's what we'd encounter in the
case of a polarized source map, or some overly simple simulations.

The "wrong_definition" argument is removed from demod.make_map.

To support HWP, the projection matrix accepts an hwp: bool argument,
which leads to it reflecting all detector gamma angles.  Applying this
reflection makes the existing projection routines easily applicable to
naively demodulated Q U timestreams.