Remove unneeded query of samples from metadata
The LoadContext operator parses the obsdb directly and gets the number of samples
per observation.  There was a stale check in the code which attempted to also get
the number of samples from the result of Context.get_meta().  However for LAT
data the sample axis does not exist in that return value, which causes an error.