Add scripts that support SO jobs compatible with toast_run (#1419)
* Add scripts that support SO jobs compatible with toast_run

Toast supports a more flexible way of running jobs where all steps
are specified in config files as a sequence of nested Pipeline
operators.  This work adds some new scripts for using this mode
of running.

- For purely synthetic observing schedules, add a script which
  writes a simulated SO instrument model to disk in a compatible
  toast format.

- Add a so-toast-run entrypoint which can generate and run per
  observation and per wafer-observation tasks with toast_run. The
  output for each task is redirected to a log file in a per-task
  directory.  Add unit test for this script which runs a simulation

- Fix simulated focalplane roundtrip to disk- numpy scalars were
  being dumped instead of python float.

- Fix some unrelated unit tests which were not being skipped in
  MPI jobs.

* Compute the pseudo-cls from the coadded map

* When coadding wafer-observation maps, use the low/mid ell (noise dominated) <BB>
as a proxy for the noise level in our signal band of ell space.  Cut outliers
based on this and optionally compute an additional noise weighting that
captures the excess noise in this ell range relative to the mid/high ell values
which should be well represented by the pixel noise covariance.

* Small cleanups to the plots