Added in reindex_axis and handling for tdets
Added in reindex_axis function in the AxisManager class which allows one to reindex all datasets that are assigned to a specified axis in an AxisManager. This is useful for dealing with tdets in get_obs.

Also added in handling for tdets in get_obs. The handling merges the dets+tdets axis, inserts the tdets signal into the dets signal dataset, and reindexes every dataset in the obs_aman that is assigned to the dets axis (det_info... etc.). These axes will have nans inserted where the tdets exist in the new merged dets axis.